Rename payment to payments so it matches telegrams own naming
