+++ telegram/__init__.py
@@ -48,10 +48,10 @@ from .parsemode import ParseMode
 from .messageentity import MessageEntity
 from .games.animation import Animation
 from .games.game import Game
-from .payment.shippingaddress import ShippingAddress
-from .payment.orderinfo import OrderInfo
-from .payment.successfulpayment import SuccessfulPayment
-from .payment.invoice import Invoice
+from .payments.shippingaddress import ShippingAddress
+from .payments.orderinfo import OrderInfo
+from .payments.successfulpayment import SuccessfulPayment
+from .payments.invoice import Invoice
 from .message import Message
 from .callbackquery import CallbackQuery
 from .choseninlineresult import ChosenInlineResult
@@ -84,10 +84,10 @@ from .inline.inputtextmessagecontent import InputTextMessageContent
 from .inline.inputlocationmessagecontent import InputLocationMessageContent
 from .inline.inputvenuemessagecontent import InputVenueMessageContent
 from .inline.inputcontactmessagecontent import InputContactMessageContent
-from .payment.labeledprice import LabeledPrice
-from .payment.shippingoption import ShippingOption
-from .payment.precheckoutquery import PreCheckoutQuery
-from .payment.shippingquery import ShippingQuery
+from .payments.labeledprice import LabeledPrice
+from .payments.shippingoption import ShippingOption
+from .payments.precheckoutquery import PreCheckoutQuery
+from .payments.shippingquery import ShippingQuery
 from .webhookinfo import WebhookInfo
 from .games.gamehighscore import GameHighScore
 from .update import Update
