+++ telegram/payments/shippingquery.py
